// src/generated/UsageStatisticsAttributionDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type UsageStatisticsAttributionDto = { key: string, label: string, isOther: boolean, value: string, sharePercent: string, redirectedRequestCount: number, redirectedTotalTokens: string, };