// src/generated/UsageStatisticsBucketDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UsageStatisticsTokensDto } from "./UsageStatisticsTokensDto";

export type UsageStatisticsBucketDto = { startedAtMs: number, finishedAtMs: number, label: string, requestCount: number, tokens: UsageStatisticsTokensDto, costPicoUsd: string, };